# docs/openapi/appointments-service.yaml
openapi: 3.1.0
info:
  title: Appointments Service
  version: 1.0.0
  description: Agenda médica, slots de disponibilidad y ciclo de vida de citas.
  contact:
    email: cra2028252@est.univalle.edu

servers:
  - url: http://localhost/api/appointments
    description: Vía Nginx → Kong (local)
  - url: http://localhost:8000/api/appointments
    description: Directo a Kong (local)
  - url: http://kong:8000/api/appointments
    description: Directo a Kong desde la red Docker (uso en pruebas)

tags:
  - name: Slots
    description: Disponibilidad de médicos
  - name: Appointments
    description: Gestión de citas
  - name: Observability
    description: Health y métricas

paths:
  /slots:
    get:
      description: Listar slots
      operationId: listSlots
      tags: [Slots]
      parameters:
        - in: query
          name: doctorId
          schema:
            type: string
            format: uuid
          description: Filtrar por médico
        - in: query
          name: available
          schema:
            type: boolean
            default: true
          description: Solo slots disponibles
      responses:
        '200':
          description: Lista de slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Slot'

    post:
      description: Crear slot
      operationId: createSlot
      tags: [Slots]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSlotRequest'
            example:
              doctorId: 00000000-0000-0000-0000-000000000001
              startsAt: "2026-06-10T10:00:00Z"
              endsAt:   "2026-06-10T10:30:00Z"
      responses:
        '201':
          description: Slot creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Slot'
              examples:
                created:
                  summary: Ejemplo de slot creado
                  value:
                    id: 75e0557a-8037-4352-83b3-b860d5d37a9b
                    doctorId: 00000000-0000-0000-0000-000000000001
                    startsAt: "2026-06-10T10:00:00Z"
                    endsAt: "2026-06-10T10:30:00Z"
                    isAvailable: true
        '400':
          $ref: '#/components/responses/BadRequest'

  /:
    post:
      description: Crear cita
      operationId: createAppointment
      tags: [Appointments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentRequest'
            example:
              patientId: 00000000-0000-0000-0000-000000000002
              doctorId:  00000000-0000-0000-0000-000000000001
              slotId:    00000000-0000-0000-0000-000000000003
              notes: Primera consulta
      responses:
        '201':
          description: Cita creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
              examples:
                created:
                  summary: Ejemplo de cita creada
                  value:
                    id: f78d4bcb-9425-467f-a4f6-5e26c15cff6a
                    patientId: 00000000-0000-0000-0000-000000000002
                    doctorId: 00000000-0000-0000-0000-000000000001
                    slotId: 00000000-0000-0000-0000-000000000003
                    startsAt: "2026-06-10T10:00:00Z"
                    endsAt: "2026-06-10T10:30:00Z"
                    status: pending
                    notes: Primera consulta
                    createdAt: "2026-06-10T12:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /{id}:
    parameters:
      - $ref: '#/components/parameters/UUIDPath'
    get:
      description: Obtener cita por ID
      operationId: getAppointment
      tags: [Appointments]
      responses:
        '200':
          description: Cita encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '404':
          $ref: '#/components/responses/NotFound'

  /{id}/confirm:
    parameters:
      - $ref: '#/components/parameters/UUIDPath'
    patch:
      description: Confirmar cita
      operationId: confirmAppointment
      tags: [Appointments]
      responses:
        '200':
          description: Cita confirmada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '404':
          $ref: '#/components/responses/NotFound'

  /{id}/reschedule:
    parameters:
      - $ref: '#/components/parameters/UUIDPath'
    patch:
      description: Reagendar cita
      operationId: rescheduleAppointment
      tags: [Appointments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RescheduleRequest'
            example:
              slotId: 00000000-0000-0000-0000-000000000004
      responses:
        '200':
          description: Cita reagendada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /{id}/cancel:
    parameters:
      - $ref: '#/components/parameters/UUIDPath'
    patch:
      description: Cancelar cita
      operationId: cancelAppointment
      tags: [Appointments]
      responses:
        '200':
          description: Cita cancelada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '404':
          $ref: '#/components/responses/NotFound'

  /patients/{patientId}:
    get:
      description: Citas de un paciente
      operationId: listByPatient
      tags: [Appointments]
      parameters:
        - in: path
          name: patientId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de citas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'

  /doctors/{doctorId}:
    get:
      description: Citas de un médico
      operationId: listByDoctor
      tags: [Appointments]
      parameters:
        - in: path
          name: doctorId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de citas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'

  /actuator/health:
    get:
      description: Health check
      operationId: health
      tags: [Observability]
      responses:
        '200':
          description: Servicio saludable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              examples:
                ok:
                  summary: Health OK
                  value:
                    status: UP

  /actuator/prometheus:
    get:
      description: Métricas Prometheus
      operationId: metrics
      tags: [Observability]
      responses:
        '200':
          description: Métricas en formato text/plain
          content:
            text/plain:
              schema:
                type: string

components:
  parameters:
    UUIDPath:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid

  responses:
    BadRequest:
      description: Datos inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Slot no disponible
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    CreateSlotRequest:
      type: object
      required: [doctorId, startsAt, endsAt]
      properties:
        doctorId:
          type: string
          format: uuid
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time

    Slot:
      type: object
      properties:
        id:
          type: string
          format: uuid
        doctorId:
          type: string
          format: uuid
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        isAvailable:
          type: boolean
      example:
        id: 75e0557a-8037-4352-83b3-b860d5d37a9b
        doctorId: 00000000-0000-0000-0000-000000000001
        startsAt: "2026-06-10T10:00:00Z"
        endsAt: "2026-06-10T10:30:00Z"
        isAvailable: true

    CreateAppointmentRequest:
      type: object
      required: [patientId, doctorId, slotId]
      properties:
        patientId:
          type: string
          format: uuid
        doctorId:
          type: string
          format: uuid
        slotId:
          type: string
          format: uuid
        notes:
          type: string

    RescheduleRequest:
      type: object
      required: [slotId]
      properties:
        slotId:
          type: string
          format: uuid

    Appointment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        doctorId:
          type: string
          format: uuid
        slotId:
          type: string
          format: uuid
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, confirmed, rescheduled, cancelled]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      example:
        id: f78d4bcb-9425-467f-a4f6-5e26c15cff6a
        patientId: 22222222-2222-2222-2222-222222222222
        doctorId: 11111111-1111-1111-1111-111111111111
        slotId: 5f5bfbf4-7833-4e13-8068-a3b14a1ed283
        startsAt: "2026-05-24T18:00:00Z"
        endsAt: "2026-05-24T18:30:00Z"
        status: cancelled
        notes: prueba
        createdAt: "2026-05-24T20:20:54.758891Z"

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: UP

    Error:
      type: object
      properties:
        error:
          type: string